requirements.txt
pymongo
gnews
python-dotenv
openai
nltk
newspaper3k
lxml
lxml.html_clean
requests
re
smtplib
unicodedata